ubi-hardening-extras/tests/requirements.txt
# Copyright 2023 The MathWorks, Inc.

docker>=6.1.2
pytest-testinfra>=9.0
